codegen: add constant definitions from value tables
Implements basic `pub const` declarations within messages for each
value associated with a signal.  Some sanitization is performed to
constrain the names to alphanumeric with underscores, though it is
possible that some DBCs will let invalid identifiers slip through
so this may need more work.  Also, the identifers are produced with
a compile-time downcast from `usize` which is not the cleanest
solution and could also be improved.

Publishing 0.1.6 release.